添加了查看下一抽概率的界面
白卡下一抽的算法存在BUG，需要改进

+++ app/src/main/java/club/quan9/luck_counter/SomePrivateClasses/ArrayAndString.java
@@ -34,7 +34,7 @@ public class ArrayAndString
     {
         double[] doublea=new double[values.length];
         for(int i=0;i<values.length;i++)
-            doublea[i]=Integer.valueOf(values[i]);
+            doublea[i]=Double.valueOf(values[i]);
         return doublea;
     }
 }
